Revert "builder design update"
This reverts commit 15d93cf030ccfb2d00aee30a0f4ae321f7ba5018.
