+++ apps/frontend/package.json
@@ -7,20 +7,17 @@
     "start": "vinxi start"
   },
   "dependencies": {
-    "@hope-ui/solid": "^0.6.7",
     "@kobalte/core": "^0.13.6",
     "@rekajs/core": "^0.1.30",
     "@rekajs/types": "^0.1.23",
     "@solidjs/router": "^0.14.1",
     "@solidjs/start": "^1.0.6",
-    "@stitches/core": "^1.2.8",
     "autoprefixer": "^10.4.19",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "postcss": "^8.4.38",
     "pretty-print-json": "^3.0.2",
     "solid-js": "^1.8.18",
-    "solid-transition-group": "^0.2.3",
     "tailwind-merge": "^2.5.2",
     "tailwindcss": "^3.4.10",
     "tailwindcss-animate": "^1.0.7",
