--- apps/frontend/src/components/active-section/index.tsx
@@ -1,1 +1,0 @@
-export * from "./active-section"
