--- apps/frontend/src/components/block-tools/index.tsx
@@ -1,2 +1,0 @@
-export * from "./block-tools"
-export { default } from "./block-tools"
