--- apps/frontend/src/components/components-container/index.tsx
@@ -1,1 +1,0 @@
-export * from "./components-container"
