--- apps/frontend/src/components/editing-panel/editing-panel.tsx
@@ -1,8 +1,0 @@
-import { JsonState } from '../dev-only/json-state/json-state';
-export const EditingPanelView = () => {
-    return (
-        <div class="w-[422px] bg-white h-full">
-            <JsonState />
-        </div>
-    )
-}
