--- apps/frontend/src/components/editing-panel/index.tsx
@@ -1,1 +1,0 @@
-export * from "./editing-panel"
