+++ apps/frontend/src/components/editor/editor.tsx
@@ -38,7 +38,7 @@ export const EditorView = () => {
   const [show, setShow] = createSignal(false);
 
   const frame = reka.createFrame({
-    id: "first-app-2",
+    id: "first-app",
     component: {
       name: "App",
       external: undefined,
@@ -56,7 +56,7 @@ export const EditorView = () => {
   });
 
   return (
-    <div class="w-full p-5 rounded-sm bg-white mt-5 overflow-auto">
+    <div class="w-full p-5">
       <p>Please press d to update the editor.</p>
       {show() && (
         <div class="w-fit mx-auto">
