--- apps/frontend/src/resources/index.tsx
@@ -1,1 +1,0 @@
-export * from "./icons"
