--- apps/frontend/src/resources/types.ts
@@ -1,7 +1,0 @@
-import { Accessor } from "solid-js"
-
-export type SidebarIconType = {
-    styles?: string,
-    activeLink: Accessor<string>
-    page: string
-}
